// client/src/utils/Constants.ts
export const GAME_CONFIG = {
  WIDTH: 1024,
  HEIGHT: 768,
  TILE_SIZE: 32,
  FPS: 60,
  PHYSICS_FPS: 60,
  BACKGROUND_COLOR: 0x1a1a2e,
  DEBUG: false
} as const;

// Networking Constants
export const NETWORK_CONFIG = {
  SERVER_URL: 'ws://localhost:3001',
  RECONNECT_ATTEMPTS: 5,
  RECONNECT_DELAY: 1000,
  PING_INTERVAL: 5000,
  TIMEOUT: 10000,
  MAX_PREDICTION_FRAMES: 60
} as const;

// Gameplay Constants
export const GAMEPLAY_CONFIG = {
  MAX_PLAYERS: 6,
  MIN_PLAYERS: 2,
  MAX_KEYS_PER_PLAYER: 3,
  BATTLE_PROXIMITY_RANGE: 2.0, // tiles
  BATTLE_PROXIMITY_TIME: 3000, // ms
  BATTLE_COUNTDOWN_TIME: 3000, // ms
  BATTLE_SELECTION_TIME: 5000, // ms
  BATTLE_COOLDOWN_TIME: 10000, // ms
  MOVEMENT_SPEED: 128, // pixels per second
  SPAWN_PROTECTION_TIME: 5000, // ms
  MATCH_DURATION: 300000 // 5 minutes in ms
} as const;
// ...
export const COLORS = {
  PRIMARY: 0x1a1a2e,
  SECONDARY: 0x16213e,
  ACCENT_CYAN: 0x00f5ff,
  ACCENT_ORANGE: 0xff6b35,
  ACCENT_PINK: 0xff006e,
  WHITE: 0xf0f0f0,
  GRAY: 0x333333,
  BLACK: 0x000000,
  
  // Player Colors
  PLAYER_COLORS: [
    0x00f5ff, // Cyan
    0xff6b35, // Orange
    0xff006e, // Pink
    0x00ff87, // Green
    0xffaa00, // Yellow
    0x8000ff  // Purple
  ],
  
  // RPS Symbol Colors
  ROCK_COLOR: 0x8b4513,
  PAPER_COLOR: 0xffffff,
  SCISSORS_COLOR: 0xc0c0c0
} as const;

// Game Enums
export enum GamePhase {
  WAITING = 'waiting',
  STARTING = 'starting',
  PLAYING = 'playing',
  ENDING = 'ending',
  FINISHED = 'finished'
}

export enum PlayerStatus {
  ALIVE = 'alive',
  ELIMINATED = 'eliminated',
  DISCONNECTED = 'disconnected',
  SPECTATING = 'spectating'
}

export enum KeyType {
  ROCK = 'rock',
  PAPER = 'paper',
  SCISSORS = 'scissors'
}
// ...
export const SCENE_KEYS = {
  BOOT: 'BootScene',
  MENU: 'MenuScene',
  GAME: 'GameScene',
  BATTLE: 'BattleScene',
  UI: 'UIScene'
} as const;
// ...
export const DEBUG_CONFIG = {
  SHOW_FPS: true,
  SHOW_COLLISION_BOXES: false,
  SHOW_GRID: false,
  SHOW_PROXIMITY_RADIUS: false,
  LOG_NETWORK_MESSAGES: true,
  LOG_GAME_EVENTS: true,
  MOCK_MULTIPLAYER: true // ENABLED for standalone testing
} as const;